// src/store/user/user.types.ts
export enum USER_ACTION_TYPES {
    SET_CURRENT_USER = 'user/SET_CURRENT_USER',
    CHECK_USER_SESSION = 'user/CHECK_USER_SESSION',
    GOOGLE_SIGN_IN_START = 'user/GOOGLE_SIGN_IN_START',
    EMAIL_SIGN_IN_START = 'user/EMAIL_SIGN_IN_START',
    USER_SIGN_IN_SUCCESS = 'user/USER_SIGN_IN_SUCCESS', 
    USER_SIGN_IN_FAILED = 'user/USER_SIGN_IN_FAILED',
    USER_SIGN_UP_START = 'user/USER_SIGN_UP_START',
    USER_SIGN_UP_SUCCESS = 'user/USER_SIGN_UP_SUCCESS',
    USER_SIGN_UP_FAILED = 'user/USER_SIGN_UP_FAILED',
    USER_SIGN_OUT_START = 'user/USER_SIGN_OUT_START',
    USER_SIGN_OUT_SUCCESS = 'user/USER_SIGN_OUT_SUCCESS',
    USER_SIGN_OUT_FAILED = 'user/USER_SIGN_OUT_FAILED',
}